Moving the glacier extent extraction to the behaviour class.
